Moving logic from RetrofitCacheAdapter to RxCacheInterceptor

+++ cache_interceptor/src/main/java/uk/co/glass_software/android/cache_interceptor/interceptors/cache/CacheManager.java
@@ -29,7 +29,7 @@ class CacheManager {
                  Function<Long, Date> dateFactory,
                  Gson gson,
                  Logger logger,
-                 long timeToLiveInMinutes) {
+                 int timeToLiveInMinutes) {
         this.databaseManager = databaseManager;
         this.dateFactory = dateFactory;
         this.gson = gson;
